Implement arbitrary argument pass-through for plans
The code will ensure that an error is raised if a plan definition in the
hierarchy has a `args` field which is not an array. However, the `args`
field in the plan definitions is optional to make the change completely
backwards-compatible.

Additionally, the pushgateway address is not passed in the Plan's
`command` field anymore, but is merged with the provided arbitrary
arguments in the `Plan()` function of the helper library.